+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Italian translations for interface
 - Welcome and Tutorial windows on first run and in About window
 - Dutch translations for interface
+- DND functionality for Mac & Windows users preventing breaks from happening while in Do Not Disturb mode.
+- user can change the default monochrome tray icon to colorful
 
 ## [0.17.0] - 2018-08-11
 - Tray menu displays the time and type of next break
@@ -17,10 +19,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - User pause will no longer be removed upon suspend/resume
 - User pause time will be corrected upon suspend/resume for the duration
   of system sleep
+- Style padding inserted for settings page 3
 
 ### Changed
 - Czech translations updated
 - Slovak translations updated
+- default tray icon is monochrome
 
 ## [0.17.0] - 2018-05-06
 ### Added
@@ -38,6 +42,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.16.0] - 2018-03-17
 ### Added
 - Russian translations for interface
+- Ability to pause until morning
 
 ### Changed
 - upper limit of microbreak duration set to 60 seconds